[HW-5]: 1. Для endpoint созданы requests и responses 2. Сделана валидация данных для созданных requests 3. Добавлены Exceptions

+++ src/Application/Services/CompanyService.cs
@@ -1,8 +1,11 @@
 using Application.DTOs.Mappings;
+using Application.Exceptions;
 using AutoMapper;
 using Domain.Entities;
 using Infrastructure.Repositories.CompanyRepository;
 using Infrastructure.Repositories.UserRepository;
+using Application.Requests;
+using Application.Responses;
 
 namespace Application.Services;
 
@@ -19,49 +22,59 @@ public class CompanyService : ICompanyService
         _userRepository = userRepository;
     }
 
-    public async Task<Company?> Add(CompanyDto companyDto)
+    public async Task<Company?> Add(CreateCompanyRequest request)
     {
-        var mappedCompany = _mapper.Map<Company>(companyDto);
-        if (mappedCompany == null)
+        var campany = new Company()
         {
-            return null;
-        }
+            CompanyName = request.CompanyName
+        };
 
-        await _companyRepository.CreateCompany(mappedCompany);
-        return mappedCompany;
+        await _companyRepository.CreateCompany(campany);
+        return campany;
     }
 
-    public async Task<bool> Update(CompanyDto companyDto)
+    public async Task<bool> Update(UpdateCompanyRequest request)
     {
-        var mappedCompany = _mapper.Map<Company>(companyDto);
-        if (mappedCompany == null)
+        if (await _companyRepository.ReadByCompanyId(request.CompanyId) == null)
         {
-            return false;
+            throw new NotFoundApplicationException("Company not found");
         }
 
-        var users = companyDto.Users;
-        foreach (var user in users)
+        var company = new Company()
         {
-            if (_userRepository.ReadById(user.Id).Result != null)
-            {
-                return false;
-            }
-        }
+            Id = request.CompanyId,
+            CompanyName = request.CompanyName
+        };
 
-        return await _companyRepository.UpdateCompany(mappedCompany);
+        var users = company.Users;
+        company.Users = users;
+        var resources = company.Resources;
+        company.Resources = resources;
+
+        return await _companyRepository.UpdateCompany(company);
     }
 
-    public Task<bool> Delete(int companyId)
+    public async Task<bool> Delete(int companyId)
     {
-        return _companyRepository.DeleteCompany(companyId);
+        if (await _companyRepository.ReadByCompanyId(companyId) == null)
+        {
+            throw new NotFoundApplicationException("Company not found");
+        }
+
+        return await _companyRepository.DeleteCompany(companyId);
     }
 
     public async Task<bool> AddUserToCompany(int userId, int companyId)
     {
+        if (await _companyRepository.ReadByCompanyId(companyId) == null)
+        {
+            throw new NotFoundApplicationException("Company not found");
+        }
+
         var user = _userRepository.ReadById(userId).Result;
         if (user == null)
         {
-            return false;
+            throw new NotFoundApplicationException("User not found");
         }
 
         return await _companyRepository.AddUserToCompany(user, companyId);
@@ -69,38 +82,82 @@ public class CompanyService : ICompanyService
 
     public async Task<bool> DeleteUserFromCompany(int userId, int companyId)
     {
+        if (await _companyRepository.ReadByCompanyId(companyId) == null)
+        {
+            throw new NotFoundApplicationException("Company not found");
+        }
+
         var user = _userRepository.ReadById(userId).Result;
         if (user == null)
         {
-            return false;
+            throw new NotFoundApplicationException("User not found");
         }
 
         return await _companyRepository.RemoveUserFromCompany(user, companyId);
     }
 
-    public async Task<CompanyDto?> GetCompany(int companyId)
+    public async Task<CompanyResponse?> GetCompany(int companyId)
     {
         var company = await _companyRepository.ReadByCompanyId(companyId);
-        var mappedCompany = _mapper.Map<CompanyDto>(company);
-        return mappedCompany;
-    }
-
-    public async Task<IEnumerable<CompanyDto?>> GetAllCompanies()
-    {
-        var companies = await _companyRepository.ReadAllCompanies();
-        var mappedCompanies = companies.Select(i => _mapper.Map<CompanyDto>(i));
-        return mappedCompanies;
-    }
-
-    public async Task<IEnumerable<UserDto?>> GetCompanyUsers(int companyId)
-    {
-        var users = await _companyRepository.ReadCompanyUsers(companyId);
-        if (users == null)
+        if (company == null)
         {
-            return null;
+            throw new NotFoundApplicationException("Company not found");
         }
 
-        var mappedUsers = users.Select(i => _mapper.Map<UserDto>(i));
-        return mappedUsers;
+        var companyResponse = new CompanyResponse()
+        {
+            CompanyName = company.CompanyName
+        };
+        return companyResponse;
+    }
+
+    public async Task<IEnumerable<CompanyResponse?>> GetAllCompanies()
+    {
+        var companies = await _companyRepository.ReadAllCompanies();
+        if (companies == null || companies.Count() == 0)
+        {
+            throw new NotFoundApplicationException("Companies not found");
+        }
+
+        var companiesResponse = new List<CompanyResponse>();
+        foreach (var company in companies)
+        {
+            companiesResponse.Add(new CompanyResponse()
+            {
+                CompanyName = company.CompanyName
+            });
+        }
+
+        return companiesResponse;
+    }
+
+    public async Task<IEnumerable<UserResponse?>> GetCompanyUsers(int companyId)
+    {
+        if (await _companyRepository.ReadByCompanyId(companyId) == null)
+        {
+            throw new NotFoundApplicationException("Company not found");
+        }
+
+        var users = await _companyRepository.ReadCompanyUsers(companyId);
+        if (users == null || users.Count() == 0)
+        {
+            throw new NotFoundApplicationException("Users not found");
+        }
+
+        var usersResponse = new List<UserResponse>();
+        foreach (var user in users)
+        {
+            usersResponse.Add(new UserResponse()
+            {
+                FirstName = user.FirstName,
+                LastName = user.LastName,
+                Email = user.Email,
+                Patronymic = user.Patronymic,
+                Username = user.Username,
+                CompanyId = user.CompanyId
+            });
+        }
+
+        return usersResponse;
     }
 }
