+++ src/Application/Services/MonitoringSettingService.cs
@@ -1,7 +1,11 @@
 using Application.DTOs.Mappings;
+using Application.Exceptions;
 using AutoMapper;
 using Domain.Entities;
 using Infrastructure.Repositories.MonitoringSettingRepository;
+using Application.Requests;
+using Application.Responses;
+using Infrastructure.Repositories.ResourceRepository;
 
 namespace Application.Services;
 
@@ -9,45 +13,76 @@ public class MonitoringSettingService : IMonitoringSettingService
 {
     private readonly IMonitoringSettingRepository _monitoringSettingRepository;
     private readonly IMapper _mapper;
+    private readonly IResourceRepository _resourceRepository;
 
-    public MonitoringSettingService(IMonitoringSettingRepository monitoringSettingRepository, IMapper mapper)
+    public MonitoringSettingService(IMonitoringSettingRepository monitoringSettingRepository, IMapper mapper,
+        IResourceRepository resourceRepository)
     {
         _mapper = mapper;
         _monitoringSettingRepository = monitoringSettingRepository;
+        _resourceRepository = resourceRepository;
     }
 
-    public async Task<MonitoringSetting?> Add(MonitoringSettingDto monitoringSettingDto)
+    public async Task<MonitoringSetting?> Add(CreateMonitoringSettingRequest request)
     {
-        var mappedMonitoringSetting = _mapper.Map<MonitoringSetting>(monitoringSettingDto);
-        if (mappedMonitoringSetting != null)
+        if (await _resourceRepository.ReadByResourceId(request.ResourceId) == null)
         {
-            await _monitoringSettingRepository.CreateSetting(mappedMonitoringSetting);
-            return mappedMonitoringSetting;
+            throw new NotFoundApplicationException("Resource not found");
         }
 
-        return null;
+        var monitoringSetting = new MonitoringSetting()
+        {
+            ResourceId = request.ResourceId,
+            CheckInterval = request.CheckInterval,
+            Mode = request.Mode
+        };
+
+        await _monitoringSettingRepository.CreateSetting(monitoringSetting);
+        return monitoringSetting;
     }
 
-    public async Task<bool> Update(MonitoringSettingDto monitoringSettingDto)
+    public async Task<bool> Update(UpdateMonitoringSettingRequest request)
     {
-        var mappedMonitoringSetting = _mapper.Map<MonitoringSetting>(monitoringSettingDto);
-        if (mappedMonitoringSetting == null)
+        if (await _monitoringSettingRepository.ReadByResourceId(request.ResourceId) == null)
         {
-            return false;
+            throw new NotFoundApplicationException("MonitoringSetting not found");
         }
 
-        return await _monitoringSettingRepository.UpdateSetting(mappedMonitoringSetting);
+        var monitoringSetting = new MonitoringSetting()
+        {
+            Id = request.Id,
+            CheckInterval = request.CheckInterval,
+            Mode = request.Mode,
+            ResourceId = request.ResourceId
+        };
+
+        return await _monitoringSettingRepository.UpdateSetting(monitoringSetting);
     }
 
     public async Task<bool> Delete(int monitoringSettingId)
     {
+        if (await _monitoringSettingRepository.ReadById(monitoringSettingId) == null)
+        {
+            throw new NotFoundApplicationException("MonitoringSetting not found");    
+        }
+        
         return await _monitoringSettingRepository.DeleteSetting(monitoringSettingId);
     }
 
-    public async Task<MonitoringSettingDto?> GetMonitoringSetting(int serviceId)
+    public async Task<MonitoringSettingResponse?> GetMonitoringSetting(int serviceId)
     {
         var monitoringSetting = await _monitoringSettingRepository.ReadByResourceId(serviceId);
-        var mappedMonitoringSetting = _mapper.Map<MonitoringSettingDto>(monitoringSetting);
-        return mappedMonitoringSetting;
+        if (monitoringSetting == null)
+        {
+            throw new NotFoundApplicationException("MonitoringSetting not found");    
+        }
+        
+        var monitoringSettingResponse = new MonitoringSettingResponse()
+        {
+            CheckInterval = monitoringSetting.CheckInterval,
+            ResourceId = monitoringSetting.ResourceId,
+        };
+
+        return monitoringSettingResponse;
     }
 }
