+++ src/Application/Services/TaskService.cs
@@ -1,10 +1,14 @@
 using Application.DTOs.Mappings;
+using Application.Exceptions;
 using AutoMapper;
 using Domain.Entities;
 using Infrastructure.Repositories.CompanyRepository;
 using Infrastructure.Repositories.ResourceRepository;
 using Infrastructure.Repositories.TaskRepository;
 using Infrastructure.Repositories.UserRepository;
+using Application.Requests;
+using Application.Responses;
+using TaskStatus = Domain.Enums.TaskStatus;
 
 namespace Application.Services;
 
@@ -26,69 +30,134 @@ public class TaskService : ITaskService
         _companyRepository = companyRepository;
     }
 
-    public async Task<ServiceTask?> Add(ServiceTaskDto serviceTaskDto)
+    public async Task<ServiceTask?> Add(CreateTaskRequest request)
     {
-        if (await _resourceRepository.ReadByResourceId(serviceTaskDto.ResourceId) == null)
+        if (await _resourceRepository.ReadByResourceId(request.ResourceId) == null)
         {
-            return null;
+            throw new NotFoundApplicationException("Resource not found");
         }
 
-        var mappedTask = _mapper.Map<ServiceTask>(serviceTaskDto);
-        if (mappedTask != null)
+        if (await _userRepository.ReadById(request.AssignedUserId) == null ||
+            await _userRepository.ReadById(request.CreatedById) == null)
         {
-            await _taskRepository.CreateTask(mappedTask);
-            return mappedTask;
+            throw new NotFoundApplicationException("User not found");
         }
 
-        return null;
+        var task = new ServiceTask()
+        {
+            ResourceId = request.ResourceId,
+            Description = request.Description,
+            AssignedUserId = request.AssignedUserId,
+            CreatedById = request.CreatedById,
+            Status = request.Status
+        };
+
+        await _taskRepository.CreateTask(task);
+        return task;
     }
 
-    public async Task<bool> Update(ServiceTaskDto serviceTaskDto)
+    public async Task<bool> Update(UpdateTaskRequest request)
     {
-        var mappedTask = _mapper.Map<ServiceTask>(serviceTaskDto);
-        if (mappedTask == null)
+        var taskToUpdate = await _taskRepository.ReadTaskId(request.TaskId);
+        if (taskToUpdate == null)
         {
-            return false;
+            throw new NotFoundApplicationException("Task not found");
         }
 
-        if (await _resourceRepository.ReadByResourceId(serviceTaskDto.ResourceId) == null)
+        if (await _resourceRepository.ReadByResourceId(request.ResourceId) == null)
         {
-            return false;
+            throw new NotFoundApplicationException("Resource not found");
         }
 
-        var assignedUserToUpdate = _userRepository.ReadById(serviceTaskDto.AssignedUserId).Result;
-        if (assignedUserToUpdate == null || assignedUserToUpdate.CompanyId !=
-            _userRepository.ReadById(serviceTaskDto.AssignedUserId).Result.CompanyId)
+        var assignedUserToUpdate = await _userRepository.ReadById(request.AssignedUserId);
+        var createdUserToUpdate = await _userRepository.ReadById(request.CreatedById);
+        if (assignedUserToUpdate == null || createdUserToUpdate == null)
         {
-            return false;
+            throw new NotFoundApplicationException("User not found");
         }
 
-        return await _taskRepository.UpdateTask(mappedTask, assignedUserToUpdate);
+        if (taskToUpdate.AssignedUser.CompanyId != assignedUserToUpdate.CompanyId ||
+            taskToUpdate.CreatedBy.CompanyId != createdUserToUpdate.CompanyId)
+        {
+            throw new UserArgumentException("User is not in the company");
+        }
+
+        var task = new ServiceTask()
+        {
+            Id = request.TaskId,
+            ResourceId = request.ResourceId,
+            Description = request.Description,
+            AssignedUserId = request.AssignedUserId,
+            CreatedById = request.CreatedById,
+            CompletionTime = request.CompletionTime,
+            Status = request.Status
+        };
+
+        return await _taskRepository.UpdateTask(task, assignedUserToUpdate);
     }
 
     public async Task<bool> Delete(int serviceTaskId)
     {
+        if (await _taskRepository.ReadTaskId(serviceTaskId) == null)
+        {
+            throw new NotFoundApplicationException("Task not found");
+        }
+
         return await _taskRepository.DeleteTask(serviceTaskId);
     }
 
-    public async Task<ServiceTaskDto?> GetTask(int taskId)
+    public async Task<TaskResponse?> GetTask(int taskId)
     {
         var serviceTask = await _taskRepository.ReadTaskId(taskId);
-        var mappedTask = _mapper.Map<ServiceTaskDto>(serviceTask);
-        return mappedTask;
+        if (serviceTask == null)
+        {
+            throw new NotFoundApplicationException("Task not found");
+        }
+
+        var taskResponse = new TaskResponse()
+        {
+            AssignedUserId = serviceTask.AssignedUserId,
+            Description = serviceTask.Description,
+            StartTime = serviceTask.StartTime,
+            CompletionTime = serviceTask.CompletionTime,
+            Status = serviceTask.Status,
+            CreatedById = serviceTask.CreatedById,
+            ResourceId = serviceTask.ResourceId
+        };
+
+        return taskResponse;
     }
 
-    public async Task<IEnumerable<ServiceTaskDto?>> GetAllCompanyTasks(int companyId)
+    public async Task<IEnumerable<TaskResponse?>> GetAllCompanyTasks(int companyId)
     {
         var company = await _companyRepository.ReadByCompanyId(companyId);
         if (company == null)
         {
-            return null;
+            throw new NotFoundApplicationException("Company not found");
         }
-        
+
         var serviceTasks = await _taskRepository.ReadAllTasksCompanyId(companyId);
-        var mappedServiceTask = serviceTasks.Select(i => _mapper.Map<ServiceTaskDto>(i));
-        return mappedServiceTask;
+        if (serviceTasks.Count() == 0 || serviceTasks == null)
+        {
+            throw new NotFoundApplicationException("Tasks not found");
+        }
+
+        var tasksResponse = new List<TaskResponse>();
+        foreach (var task in serviceTasks)
+        {
+            tasksResponse.Add(new TaskResponse()
+            {
+                AssignedUserId = task.AssignedUserId,
+                Description = task.Description,
+                StartTime = task.StartTime,
+                CompletionTime = task.CompletionTime,
+                Status = task.Status,
+                CreatedById = task.CreatedById,
+                ResourceId = task.ResourceId
+            });
+        }
+
+        return tasksResponse;
     }
 
     public async Task<bool> AssignTaskToUser(int userId, int taskId)
@@ -96,7 +165,12 @@ public class TaskService : ITaskService
         var user = await _userRepository.ReadById(userId);
         if (user == null)
         {
-            return false;
+            throw new NotFoundApplicationException("User not found");
+        }
+
+        if (await _taskRepository.ReadTaskId(taskId) == null)
+        {
+            throw new NotFoundApplicationException("Task not found");
         }
 
         return await _taskRepository.AssignTaskToUser(user, taskId);
@@ -109,31 +183,86 @@ public class TaskService : ITaskService
         {
             return false;
         }
+
+        if (await _taskRepository.ReadTaskId(taskId) == null)
+        {
+            throw new NotFoundApplicationException("Task not found");
+        }
+
         return await _taskRepository.DeleteTaskToUser(user, taskId);
     }
 
     public async Task<bool> ReassignTaskToUser(int oldUserId, int newUserId, int taskId)
     {
         var newUser = _userRepository.ReadById(newUserId).Result;
-        if (newUser == null)
+        if (newUser == null || await _userRepository.ReadById(oldUserId) == null)
         {
-            return false;
+            throw new NotFoundApplicationException("User not found");
+        }
+        
+        if (await _taskRepository.ReadTaskId(taskId) == null)
+        {
+            throw new NotFoundApplicationException("Task not found");
         }
 
         return await _taskRepository.ReassignTaskToUser(oldUserId, newUser, taskId);
     }
 
-    public async Task<ServiceTaskDto?> GetTaskForUser(int userId, int taskId)
+    public async Task<TaskResponse?> GetTaskForUser(int userId, int taskId)
     {
+        if (await _userRepository.ReadById(userId) == null)
+        {
+            throw new NotFoundApplicationException("User not found");
+        }
+        
         var serviceTask = await _taskRepository.ReadTaskUser(userId, taskId);
-        var mappedTask = _mapper.Map<ServiceTaskDto>(serviceTask);
-        return mappedTask;
+        if (serviceTask == null)
+        {
+            throw new NotFoundApplicationException("Task not found");    
+        }
+        
+        var taskResponse = new TaskResponse()
+        {
+            AssignedUserId = serviceTask.AssignedUserId,
+            Description = serviceTask.Description,
+            StartTime = serviceTask.StartTime,
+            CompletionTime = serviceTask.CompletionTime,
+            Status = serviceTask.Status,
+            CreatedById = serviceTask.CreatedById,
+            ResourceId = serviceTask.ResourceId
+        };
+
+        return taskResponse;
     }
 
-    public async Task<IEnumerable<ServiceTaskDto?>> GetAllUserTasks(int userId)
+    public async Task<IEnumerable<TaskResponse?>> GetAllUserTasks(int userId)
     {
+        if (await _userRepository.ReadById(userId) == null)
+        {
+            throw new NotFoundApplicationException("User not found");
+        }
+        
         var serviceTasksUser = await _taskRepository.ReadAllUserTasks(userId);
-        var mappedTasksUser = serviceTasksUser.Select(i => _mapper.Map<ServiceTaskDto>(i));
-        return mappedTasksUser;
+        if (serviceTasksUser.Count() == 0 || serviceTasksUser == null)
+        {
+            throw new NotFoundApplicationException("Tasks not found");
+        }
+        
+        var tasksResponse = new List<TaskResponse>();
+        foreach (var task in serviceTasksUser)
+        {
+            tasksResponse.Add(new TaskResponse()
+            {
+                AssignedUserId = task.AssignedUserId,
+                Description = task.Description,
+                StartTime = task.StartTime,
+                CompletionTime = task.CompletionTime,
+                Status = task.Status,
+                CreatedById = task.CreatedById,
+                ResourceId = task.ResourceId
+            });
+        }
+
+        return tasksResponse;
     }
 }
